# .openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CWS Next App API",
    "version": "0.1.0",
    "description": "Automated API documentation for the CWS Next App. This document is generated from Zod schemas and route handler metadata.",
    "contact": {
      "name": "API Support",
      "email": "api@crossweavesourcing.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
...
  "paths": {
...
    "/api/auth/webauthn/register-options": {
      "post": {
        "operationId": "webauthnRegisterOptions",
        "summary": "Get WebAuthn registration options",
        "description": "Generates passkey registration options for the active dashboard session.",
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Registration options generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "challenge": {
                      "type": "string",
                      "description": "WebAuthn challenge string",
                      "example": "dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk"
                    },
                    "rpId": {
                      "type": "string",
                      "description": "Relying Party ID",
                      "example": "localhost"
                    },
                    "timeout": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991,
                      "description": "Challenge timeout in ms",
                      "example": 60000
                    },
                    "allowCredentials": {
                      "description": "Allowed authenticator credentials",
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "Credential ID (base64url)"
                          },
                          "type": {
                            "type": "string",
                            "const": "public-key",
                            "example": "public-key"
                          }
                        },
                        "required": [
                          "id",
                          "type"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "userVerification": {
                      "type": "string",
                      "enum": [
                        "preferred",
                        "required",
                        "discouraged"
                      ],
                      "description": "User verification requirement",
                      "example": "preferred"
                    },
                    "rp": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "id": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "name"
                      ],
                      "additionalProperties": false
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "displayName"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "challenge",
                    "rpId",
                    "timeout",
                    "userVerification"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Cannot start registration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Request blocked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/webauthn/register-verify": {
      "post": {
        "operationId": "webauthnRegisterVerify",
        "summary": "Verify WebAuthn registration response",
        "description": "Verifies a passkey registration response and stores the credential for the active dashboard user.",
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "rawId": {
                    "type": "string"
                  },
                  "response": {
                    "type": "object",
                    "properties": {
                      "attestationObject": {
                        "type": "string"
                      },
                      "clientDataJSON": {
                        "type": "string"
                      },
                      "transports": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "required": [
                      "attestationObject",
                      "clientDataJSON"
                    ],
                    "additionalProperties": false
                  },
                  "type": {
                    "type": "string",
                    "const": "public-key"
                  },
                  "clientExtensionResults": {
                    "type": "object",
                    "propertyNames": {
                      "type": "string"
                    },
                    "additionalProperties": {}
                  },
                  "authenticatorAttachment": {
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "rawId",
                  "response",
                  "type"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Passkey registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid registration response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized or expired challenge",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Request blocked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/mobile/v1/admin/categories/{id}": {
      "put": {
        "tags": [
          "Mobile Admin"
        ],
        "summary": "Update an existing category",
        "description": "Admin only. Updates an existing category.",
        "operationId": "updateCategoryMobile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Category ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "slug": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "visible": {
                    "type": "boolean"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "name",
                  "slug",
                  "description",
                  "visible"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "cws_session",
        "description": "HMAC-signed session cookie issued by the web dashboard login flow. Obtained via POST /api/auth/login (Server Action) or Google OAuth callback."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Ed25519-signed JWT issued by the mobile authentication endpoints. Obtained via POST /api/mobile/v1/auth/password or /api/mobile/v1/auth/google."
      },
      "pendingSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "cws_2fa_pending",
        "description": "Short-lived HMAC-signed pending session cookie set during MFA challenge. Valid for 5 minutes. Used by WebAuthn and TOTP verification endpoints."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "A human-readable error message",
            "example": "Resource not found"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
    }
  }
}